Replace CommonJS export in server.ts with pure ESM default export
server.ts mixed module.exports with ES module imports, which breaks
under strict ESM and confuses test runners/bundlers. server.ts now
just re-exports the app; all startup/listen/shutdown logic moves to
a new start.ts so app.ts/server.ts stay importable in tests without
side effects.
