+++ package.json
@@ -2,11 +2,11 @@
   "name": "zicket-backend",
   "version": "1.0.0",
   "description": "Backend services for Zicket",
-  "main": "server.ts",
+  "main": "start.ts",
   "scripts": {
     "build": "tsc",
-    "start": "node dist/server.js",
-    "dev": "nodemon --watch 'src/**/*.ts' --exec npx ts-node src/server.ts",
+    "start": "node dist/start.js",
+    "dev": "nodemon --watch 'src/**/*.ts' --exec npx ts-node src/start.ts",
     "test": "jest",
     "test:watch": "jest --watch",
     "format": "prettier --write .",
